increasing precision

+++ tests/src/nn_models/sequential/persist_code.cpp
@@ -23,7 +23,7 @@ namespace rlt = RL_TOOLS_NAMESPACE_WRAPPER ::rl_tools;
 #include <optional>
 #include <string>
 
-using T = float;
+using T = double;
 using DEVICE = rlt::devices::DefaultCPU;
 using TI = typename DEVICE::index_t;
